+++ PlexMatchGenerator/Constants/MediaConstants.cs
@@ -5,5 +5,6 @@
         public const string PlexMatchTitleHeader = "Title: ";
         public const string PlexMatchYearHeader = "Year: ";
         public const string PlexMatchGuidHeader = "Guid: ";
+        public const string PlexMatchSeasonHeader = "Season: ";
     }
 }