+++ PlexMatchGenerator/Models/PlexMatchFileType.cs
@@ -1,0 +1,8 @@
+﻿namespace PlexMatchGenerator.Models
+{
+    public enum PlexMatchFileType
+    {
+        Main,
+        Season
+    }
+}