+++ PlexMatchGenerator/Options/GeneratorOptions.cs
@@ -10,5 +10,6 @@
         public int ItemsPerPage { get; set; }
         public List<string> LibraryNames { get; set; }
         public List<string> ShowNames { get; set; }
+        public bool EnablePerSeasonProcessing { get; set; }
     }
 }