+++ PlexMatchGenerator/RestModels/ShowOrdering.cs
@@ -1,0 +1,11 @@
+﻿namespace PlexMatchGenerator.RestModels
+{
+    public enum ShowOrdering
+    {
+        Default,
+        TMDBAired,
+        TVDBAired,
+        TVDBDVD,
+        TVDBAbsolute
+    }
+}