+++ .github/workflows/ReleaseNotes.md
@@ -3,7 +3,8 @@
 ## What's Changed
 
 * Add library command line switch to restrict to only specified libraries per [#38](https://github.com/johnkiddjr/PlexMatch-File-Generator/issues/38)
-* Add library command line switch to restrict to only specified media items per [#37](https://github.com/johnkiddjr/PlexMatch-File-Generator/issues/37)
+* Add show command line switch to restrict to only specified media items per [#37](https://github.com/johnkiddjr/PlexMatch-File-Generator/issues/37)
+* Add seasonprocessing command line switch to enable per season processing of TV Shows per [#39](https://github.com/johnkiddjr/PlexMatch-File-Generator/issues/39)
 
 ## Known Bugs